refac: move placeholder data to svele compoennt
